+++ src/test/java/test/validation/Cartesian.java
@@ -7,9 +7,8 @@ import java.util.Iterator;
 import be.baur.sda.DataNode;
 import be.baur.sda.SDA;
 import be.baur.sds.SDS;
-import be.baur.sds.validation.Error;
-import be.baur.sds.validation.ErrorList;
 import be.baur.sds.validation.Validator;
+import be.baur.sds.validation.Validator.Errors;
 import test.Test;
 
 public final class Cartesian {
@@ -29,9 +28,9 @@ public final class Cartesian {
 		InputStream sds = Cartesian.class.getResourceAsStream("/cartesian.sds");
 		Validator validator = SDS.parse(new InputStreamReader(sds, "UTF-8")).newValidator();
 
-		ErrorList errors = validator.validate(document);
+		Errors errors = validator.validate(document);
 		//for (Error error : errors) System.out.println(error.toString());
-		Iterator<Error> e = errors.iterator();
+		Iterator<?> e = errors.iterator();
 		
 		t.ts1("F01", e.next() + "", "/cartesian/line[5]/point[1]: content missing at end of 'point'; expected 'y'");
 		t.ts1("F02", e.next() + "", "/cartesian/line[5]/point[2]: content missing at end of 'point'; expected 'y'");